+++ src/main/java/Application.java
@@ -1,4 +1,4 @@
-import controller.BlackJackController;
+import blackjack.controller.BlackJackController;
 
 public class Application {
     public static void main(String[] args) {
