+++ src/main/java/blackjack/domain/game/MatchResult.java
@@ -1,0 +1,17 @@
+package blackjack.domain.game;
+
+public enum MatchResult {
+
+    WIN("승"),
+    LOSE("패");
+
+    private final String value;
+
+    MatchResult(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
