+++ src/main/java/blackjack/domain/participant/Name.java
@@ -1,0 +1,45 @@
+package blackjack.domain.participant;
+
+import java.util.Objects;
+
+public class Name {
+
+    private static final String EMPTY_NAME_ERROR = "빈 값은 이름으로 등록할 수 없습니다.";
+
+    private final String value;
+
+    public Name(final String value) {
+        validate(value);
+        this.value = value;
+    }
+
+    private void validate(String value) {
+        if (value == null || value.isBlank()) {
+            throw new IllegalArgumentException(EMPTY_NAME_ERROR);
+        }
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Name name = (Name) o;
+        return value.equals(name.value);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(value);
+    }
+
+    @Override
+    public String toString() {
+        return "Name{" +
+                "value='" + value + '\'' +
+                '}';
+    }
+}
